+++ packages/components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @devup-ui/components
 
+## 0.1.22
+
+### Patch Changes
+
+- Updated dependencies [84d0c1f]
+  - @devup-ui/react@1.0.19
+
+## 0.1.21
+
+### Patch Changes
+
+- c09438a: Fix button css
+- e1e827c: Checkbox component has been implemented.
+
 ## 0.1.20
 
 ### Patch Changes